# backend/api/server_gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package api

import (
	"fmt"
	"net/http"

	"github.com/labstack/echo/v4"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// イベントの一覧を取得します
	// (GET /events)
	GetEvents(ctx echo.Context, params GetEventsParams) error
	// 新しいイベントを作成します
	// (POST /events)
	PostEvent(ctx echo.Context) error
	// 指定したイベントを削除します
	// (DELETE /events/{eventID})
	DeleteEvent(ctx echo.Context, eventID openapi_types.UUID) error
	// 指定したイベントの詳細を取得します
	// (GET /events/{eventID})
	GetEvent(ctx echo.Context, eventID openapi_types.UUID) error
	// 指定したイベントの情報を更新します
	// (PATCH /events/{eventID})
	PatchEvent(ctx echo.Context, eventID openapi_types.UUID) error
	// 指定したイベントの出席登録をキャンセルします
	// (DELETE /events/{eventID}/attendance)
	DeleteAttendance(ctx echo.Context, eventID openapi_types.UUID) error
	// 指定したイベントに出席登録を行います
	// (POST /events/{eventID}/attendance)
	PostAttendance(ctx echo.Context, eventID openapi_types.UUID) error
	// 指定したイベントの抽選の一覧を取得します
	// (GET /events/{eventID}/lotteries)
	GetLotteries(ctx echo.Context, eventID openapi_types.UUID, params GetLotteriesParams) error
	// 指定したイベントに対して新しい抽選を作成します
	// (POST /events/{eventID}/lotteries)
	PostLottery(ctx echo.Context, eventID openapi_types.UUID) error
	// 指定した抽選を削除します
	// (DELETE /events/{eventID}/lotteries/{lotteryID})
	DeleteLottery(ctx echo.Context, eventID openapi_types.UUID, lotteryID openapi_types.UUID) error
	// 指定した抽選の詳細を取得します
	// (GET /events/{eventID}/lotteries/{lotteryID})
	GetLottery(ctx echo.Context, eventID openapi_types.UUID, lotteryID openapi_types.UUID) error
	// 指定した抽選を実行し、当選者を決定します
	// (POST /events/{eventID}/lotteries/{lotteryID})
	RollLottery(ctx echo.Context, eventID openapi_types.UUID, lotteryID openapi_types.UUID, params RollLotteryParams) error
	// 自分のtraQ IDを取得します
	// (GET /users/me)
	GetMe(ctx echo.Context) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// GetEvents converts echo context to params.
func (w *ServerInterfaceWrapper) GetEvents(ctx echo.Context) error {
	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetEventsParams
	// ------------- Required query parameter "ifDeleted" -------------

	err = runtime.BindQueryParameter("form", true, true, "ifDeleted", ctx.QueryParams(), &params.IfDeleted)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter ifDeleted: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetEvents(ctx, params)
	return err
}

// PostEvent converts echo context to params.
func (w *ServerInterfaceWrapper) PostEvent(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PostEvent(ctx)
	return err
}

// DeleteEvent converts echo context to params.
func (w *ServerInterfaceWrapper) DeleteEvent(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "eventID" -------------
	var eventID openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "eventID", ctx.Param("eventID"), &eventID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter eventID: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.DeleteEvent(ctx, eventID)
	return err
}

// GetEvent converts echo context to params.
func (w *ServerInterfaceWrapper) GetEvent(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "eventID" -------------
	var eventID openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "eventID", ctx.Param("eventID"), &eventID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter eventID: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetEvent(ctx, eventID)
	return err
}

// PatchEvent converts echo context to params.
func (w *ServerInterfaceWrapper) PatchEvent(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "eventID" -------------
	var eventID openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "eventID", ctx.Param("eventID"), &eventID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter eventID: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PatchEvent(ctx, eventID)
	return err
}

// DeleteAttendance converts echo context to params.
func (w *ServerInterfaceWrapper) DeleteAttendance(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "eventID" -------------
	var eventID openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "eventID", ctx.Param("eventID"), &eventID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter eventID: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.DeleteAttendance(ctx, eventID)
	return err
}

// PostAttendance converts echo context to params.
func (w *ServerInterfaceWrapper) PostAttendance(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "eventID" -------------
	var eventID openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "eventID", ctx.Param("eventID"), &eventID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter eventID: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PostAttendance(ctx, eventID)
	return err
}

// GetLotteries converts echo context to params.
func (w *ServerInterfaceWrapper) GetLotteries(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "eventID" -------------
	var eventID openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "eventID", ctx.Param("eventID"), &eventID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter eventID: %s", err))
	}

	// Parameter object where we will unmarshal all parameters from the context
	var params GetLotteriesParams
	// ------------- Required query parameter "ifDeleted" -------------

	err = runtime.BindQueryParameter("form", true, true, "ifDeleted", ctx.QueryParams(), &params.IfDeleted)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter ifDeleted: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetLotteries(ctx, eventID, params)
	return err
}

// PostLottery converts echo context to params.
func (w *ServerInterfaceWrapper) PostLottery(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "eventID" -------------
	var eventID openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "eventID", ctx.Param("eventID"), &eventID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter eventID: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PostLottery(ctx, eventID)
	return err
}

// DeleteLottery converts echo context to params.
func (w *ServerInterfaceWrapper) DeleteLottery(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "eventID" -------------
	var eventID openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "eventID", ctx.Param("eventID"), &eventID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter eventID: %s", err))
	}

	// ------------- Path parameter "lotteryID" -------------
	var lotteryID openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "lotteryID", ctx.Param("lotteryID"), &lotteryID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter lotteryID: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.DeleteLottery(ctx, eventID, lotteryID)
	return err
}

// GetLottery converts echo context to params.
func (w *ServerInterfaceWrapper) GetLottery(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "eventID" -------------
	var eventID openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "eventID", ctx.Param("eventID"), &eventID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter eventID: %s", err))
	}

	// ------------- Path parameter "lotteryID" -------------
	var lotteryID openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "lotteryID", ctx.Param("lotteryID"), &lotteryID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter lotteryID: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetLottery(ctx, eventID, lotteryID)
	return err
}

// RollLottery converts echo context to params.
func (w *ServerInterfaceWrapper) RollLottery(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "eventID" -------------
	var eventID openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "eventID", ctx.Param("eventID"), &eventID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter eventID: %s", err))
	}

	// ------------- Path parameter "lotteryID" -------------
	var lotteryID openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "lotteryID", ctx.Param("lotteryID"), &lotteryID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter lotteryID: %s", err))
	}

	// Parameter object where we will unmarshal all parameters from the context
	var params RollLotteryParams
	// ------------- Required query parameter "ifDuplicated" -------------

	err = runtime.BindQueryParameter("form", true, true, "ifDuplicated", ctx.QueryParams(), &params.IfDuplicated)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter ifDuplicated: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.RollLottery(ctx, eventID, lotteryID, params)
	return err
}

// GetMe converts echo context to params.
func (w *ServerInterfaceWrapper) GetMe(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetMe(ctx)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.GET(baseURL+"/events", wrapper.GetEvents)
	router.POST(baseURL+"/events", wrapper.PostEvent)
	router.DELETE(baseURL+"/events/:eventID", wrapper.DeleteEvent)
	router.GET(baseURL+"/events/:eventID", wrapper.GetEvent)
	router.PATCH(baseURL+"/events/:eventID", wrapper.PatchEvent)
	router.DELETE(baseURL+"/events/:eventID/attendance", wrapper.DeleteAttendance)
	router.POST(baseURL+"/events/:eventID/attendance", wrapper.PostAttendance)
	router.GET(baseURL+"/events/:eventID/lotteries", wrapper.GetLotteries)
	router.POST(baseURL+"/events/:eventID/lotteries", wrapper.PostLottery)
	router.DELETE(baseURL+"/events/:eventID/lotteries/:lotteryID", wrapper.DeleteLottery)
	router.GET(baseURL+"/events/:eventID/lotteries/:lotteryID", wrapper.GetLottery)
	router.POST(baseURL+"/events/:eventID/lotteries/:lotteryID", wrapper.RollLottery)
	router.GET(baseURL+"/users/me", wrapper.GetMe)

}