# docs/api.yaml
openapi: 3.0.4
info:
  title: API
  version: 0.0.0
servers:
  - url: http://localhost:8080
    description: Local Server
  - url: https://traquji.trap.show
    description: Production Server
tags:
  - name: User
  - name: Event
  - name: Attendance
  - name: Lottery
paths:
  /users/me:
    get:
      tags:
        - User
      operationId: getMe
      summary: 自分のtraQ IDを取得します
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
        '500':
          description: Internal Server Error
  /events:
    post:
      tags:
        - Event
      operationId: postEvent
      summary: 新しいイベントを作成します
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventBase'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  event_id:
                    type: string
                    format: uuid
        '500':
          description: Internal Server Error
    get:
      tags:
        - Event
      operationId: getEvents
      summary: イベントの一覧を取得します
      parameters:
        - in: query
          name: ifDeleted
          required: true
          schema:
            type: boolean
          description: このパラメータが true の場合、削除済みのイベントも含めて取得します
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventSummary'
        '500':
          description: Internal Server Error
  /events/{eventID}:
    get:
      tags:
        - Event
      operationId: getEvent
      summary: 指定したイベントの詳細を取得します
      parameters:
        - in: path
          name: eventID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '500':
          description: Internal Server Error
    patch:
      tags:
        - Event
      operationId: patchEvent
      summary: 指定したイベントの情報を更新します
      parameters:
        - in: path
          name: eventID
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventOnUpdate'
      responses:
        '200':
          description: OK
        '500':
          description: Internal Server Error
    delete:
      tags:
        - Event
      operationId: deleteEvent
      summary: 指定したイベントを削除します
      parameters:
        - in: path
          name: eventID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
        '500':
          description: Internal Server Error
  /events/{eventID}/attendance:
    post:
      tags:
        - Attendance
      operationId: postAttendance
      summary: 指定したイベントに出席登録を行います
      parameters:
        - in: path
          name: eventID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
        '500':
          description: Internal Server Error
    delete:
      tags:
        - Attendance
      operationId: deleteAttendance
      summary: 指定したイベントの出席登録をキャンセルします
      parameters:
        - in: path
          name: eventID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
        '500':
          description: Internal Server Error
  /events/{eventID}/lotteries:
    post:
      tags:
        - Lottery
      operationId: postLottery
      summary: 指定したイベントに対して新しい抽選を作成します
      parameters:
        - in: path
          name: eventID
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
              required:
                - title
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  lottery_id:
                    type: string
                    format: uuid
        '500':
          description: Internal Server Error
    get:
      tags:
        - Lottery
      operationId: getLotteries
      summary: 指定したイベントの抽選の一覧を取得します
      parameters:
        - in: path
          name: eventID
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: ifDeleted
          required: true
          schema:
            type: boolean
          description: このパラメータが true の場合、削除済みの抽選も含めて取得します
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Lottery'
        '500':
          description: Internal Server Error
  /events/{eventID}/lotteries/{lotteryID}:
    post:
      tags:
        - Lottery
      operationId: rollLottery
      summary: 指定した抽選を実行し、当選者を決定します
      parameters:
        - in: path
          name: eventID
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: lotteryID
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: ifDuplicated
          required: true
          schema:
            type: boolean
          description: このパラメータが true の場合、同一イベント内での重複当選を許可します (false の場合も同一抽選内では重複当選は許可されません)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  winner:
                    type: string
                    description: traQ ID
        '500':
          description: Internal Server Error
    get:
      tags:
        - Lottery
      operationId: getLottery
      summary: 指定した抽選の詳細を取得します
      parameters:
        - in: path
          name: eventID
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: lotteryID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lottery'
        '500':
          description: Internal Server Error
    delete:
      tags:
        - Lottery
      operationId: deleteLottery
      summary: 指定した抽選を削除します
      parameters:
        - in: path
          name: eventID
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: lotteryID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
        '500':
          description: Internal Server Error
components:
  schemas:
    EventSummary:
      type: object
      properties:
        event_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        date:
          type: string
          format: date
        is_open:
          type: boolean
        is_me_attendee:
          type: boolean
        admins:
          type: array
          items:
            type: string
            description: traQ ID
      required:
        - event_id
        - title
        - description
        - date
        - is_open
        - is_me_attendee
        - admins
    EventBase:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        date:
          type: string
          format: date
        is_open:
          type: boolean
        is_me_attendee:
          type: boolean
        admins:
          type: array
          items:
            type: string
            description: traQ ID
        attendees:
          type: array
          items:
            type: string
            description: traQ ID
      required:
        - title
        - description
        - date
        - is_open
        - is_me_attendee
        - admins
        - attendees
    EventOnUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        date:
          type: string
          format: date
        is_open:
          type: boolean
        admins:
          type: array
          items:
            type: string
            description: traQ ID
        attendees:
          type: array
          items:
            type: string
            description: traQ ID
      required:
        - title
        - description
        - date
        - is_open
        - admins
        - attendees
    Event:
      allOf:
        - $ref: '#/components/schemas/EventBase'
        - type: object
          properties:
            event_id:
              type: string
              format: uuid
            is_deleted:
              type: boolean
            updated_at:
              type: string
              format: date-time
            created_at:
              type: string
              format: date-time
          required:
            - event_id
            - is_deleted
            - updated_at
            - created_at
    Lottery:
      type: object
      properties:
        lottery_id:
          type: string
          format: uuid
        event_id:
          type: string
          format: uuid
        title:
          type: string
        is_deleted:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        winners:
          type: array
          items:
            type: string
            description: traQ ID
      required:
        - lottery_id
        - event_id
        - title
        - is_deleted
        - updated_at
        - created_at
        - winners